Control leapp-report format
In order to be completely safe that any changes that land in leapp
reporting don't break older clients there should be a possibility
to force specific version of json schema.
Currently there are 2 schemas that are actually backward compatible,
but it may not be the rule soon.

This patch adds a --report-schema option to leapp preupgrade/upgrade
commands that force sticking to specific version of the report schema.
For example with --report-schema '1.0.0' passed there will be no stable
keys in the report messages, as those have been added in the recent
'1.1.0' report schema.