cli: Automatic command discovery for leapp CLI
This patch introduces a new approach in detecting the leapp cli commands.
The main function will now look into the directory of leapp.cli.commands
and list all directories within that directory and attempts to import a
`register` function from them. If that succeeds it will call this function
passing the toplevel command object to it allowing the function to register
the commands and if necessary sub commands to it.
... TODO

Signed-off-by: Vinzenz Feenstra <vfeenstr@redhat.com>